Fix conversion warnings

+++ demo/Application.cpp
@@ -83,7 +83,9 @@ namespace demo
     {
         renderTarget = sr::RenderTarget(width, height);
 
-        projection.setPerspective(sr::tau / 6.0F, static_cast<float>(width) / static_cast<float>(height), 1.0F, 1000.0F);
+        projection.setPerspective(static_cast<float>(sr::tau / 6.0),
+                                  static_cast<float>(width) / static_cast<float>(height),
+                                  1.0F, 1000.0F);
 
         view.setTranslation(0.0F, 0.0F, 100.0F);
     }
@@ -114,6 +116,8 @@ namespace demo
     {
         renderTarget.resize(width, height);
 
-        projection.setPerspective(sr::tau / 6.0F, static_cast<float>(width) / static_cast<float>(height), 1.0F, 1000.0F);
+        projection.setPerspective(static_cast<float>(sr::tau / 6.0),
+                                  static_cast<float>(width) / static_cast<float>(height),
+                                  1.0F, 1000.0F);
     }
 }
